% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customize_schedule.R
\name{customize_schedule}
\alias{customize_schedule}
\title{Customize cronR and taskscheduleR run dates}
\usage{
customize_schedule(
  start_date = NULL,
  specific_dates = NULL,
  days_of_week = NULL,
  biweekly = FALSE,
  skip_weekends = FALSE,
  dates_to_skip = NULL,
  run_next_day = FALSE,
  days_of_month = NULL,
  select_min = FALSE,
  select_last = FALSE,
  get_dataset = FALSE
)
}
\arguments{
\item{start_date}{A string date value of Y-m-d format. Default is 01-01 of the year.}

\item{specific_dates}{A string date or vector of dates in Y-m-d format.}

\item{days_of_week}{A string name of day or vector of days in week. Valid arguments are "Monday","Tuesday","Wednesday","Thursday","Friday","Saturday","Sunday".}

\item{biweekly}{Includes every other week dates if TRUE.}

\item{skip_weekends}{Skips Saturday and Sunday if TRUE.}

\item{dates_to_skip}{A string date or vector of dates in Y-m-d format to skip.}

\item{run_next_day}{Runs the following day after dates to skip if TRUE.}

\item{days_of_month}{An integer or vector of integers of specific days in month.}

\item{select_min}{Selects the first day of days_of_month for every month if TRUE.}

\item{select_last}{Selects the last day of the month the job should run based on other criteria.}

\item{get_dataset}{Returns dataset of dates in Global Environment if TRUE.}
}
\value{
Returns a print statement of successful run if system date is among specified conditions, otherwise returns an error that halts task execution when running cronR or taskscheduleR. Returns a data frame of run dates if get_dataset is TRUE.
}
\description{
\code{custom_schedule} returns an error if the current system date is not in the specifications defined in the vector and halts execution of the script.
}
\details{
Wrapper for cronR or taskscheduleR that provides more flexibility and convenience in customizing run dates. Version 0.1.0 requires cronR or taskscheduleR to be installed and a job or task scheduled for daily at the desired time and daily frequency-minutely, hourly, or once daily with the custom_schedule function defined at the beginning of the automated script.
}
